refactor: 베팅값 long 변경으로 인한 영향 범위 수정

+++ src/main/java/blackjack/domain/Players.java
@@ -26,7 +26,7 @@ public class Players {
         List<Player> result = new ArrayList<>();
         for (PlayerBettingRequest playerBettingRequest : playersBettingRequest.value()) {
             String nickname = playerBettingRequest.playerNickname();
-            int amount = playerBettingRequest.amount();
+            long amount = playerBettingRequest.amount();
             result.add(new Player(nickname, Role.PLAYER, amount));
         }
         return from(result);
